simplify config loading
